MET-182: refactor: move save logic into Saver
