+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/event_subscribers.yml
@@ -49,7 +49,7 @@ services:
     pim_catalog.event_subscriber.product.association.save_two_way_product_associations:
         class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Product\Association\SaveTwoWayProductAssociations
         arguments:
-            - '@database_connection'
+            - '@pim_catalog.saver.two_way_product_associations_saver'
         tags:
             - { name: kernel.event_subscriber }
