+++ tests/back/Pim/Enrichment/Specification/Bundle/EventSubscriber/Product/Association/SaveTwoWayProductAssociationsSpec.php
@@ -2,6 +2,7 @@
 
 namespace Specification\Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Product\Association;
 
+use Akeneo\Pim\Enrichment\Bundle\Doctrine\Common\Saver\TwoWayProductAssociationsSaver;
 use Akeneo\Pim\Enrichment\Component\Product\Model\Product;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductAssociation;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductAssociationInterface;
@@ -10,7 +11,6 @@ use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModel;
 use Akeneo\Pim\Structure\Component\Model\AssociationType;
 use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Doctrine\Common\Collections\Collection;
-use Doctrine\DBAL\Connection;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Symfony\Component\EventDispatcher\GenericEvent;
@@ -18,27 +18,22 @@ use Symfony\Component\EventDispatcher\GenericEvent;
 class SaveTwoWayProductAssociationsSpec extends ObjectBehavior
 {
     public function let(
-        Connection $connectionNotUsed,
-        Connection $connection,
         Product $product,
-        ProductInterface $productWithTwoWayAssociation,
-        ProductAssociationInterface $productAssociation,
-        AssociationTypeInterface $associationType,
-        Collection $associationCollection
+        ProductAssociationInterface $oneWayAssociation,
+        ProductAssociationInterface $twoWayAssociation,
+        AssociationTypeInterface $oneWayAssociationType,
+        AssociationTypeInterface $twoWayAssociationType,
+        TwoWayProductAssociationsSaver $productAssociationsSaver
     ) {
-        $connectionNotUsed->beginTransaction()->shouldNotBeCalled();
-        $connectionNotUsed->executeQuery(Argument::any())->shouldNotBeCalled();
-        $connectionNotUsed->executeUpdate(Argument::any())->shouldNotBeCalled();
-
         $product->getId()->willReturn(42);
-        $productWithTwoWayAssociation->getId()->willReturn(42);
-        $productWithTwoWayAssociation->getAssociations()->willReturn($associationCollection);
-        $associationCollection->toArray()->willReturn([$productAssociation]);
-        $productAssociation->getAssociationType()->willReturn($associationType);
-        $associationType->isTwoWay()->willReturn(true);
-        $associationType->getId()->willReturn(1);
 
-        $this->beConstructedWith($connection);
+        $twoWayAssociation->getAssociationType()->willReturn($twoWayAssociationType);
+        $twoWayAssociationType->isTwoWay()->willReturn(true);
+
+        $oneWayAssociation->getAssociationType()->willReturn($oneWayAssociationType);
+        $oneWayAssociationType->isTwoWay()->willReturn(false);
+
+        $this->beConstructedWith($productAssociationsSaver);
     }
 
     public function it_subscribe_to_the_save_event()
@@ -49,67 +44,63 @@ class SaveTwoWayProductAssociationsSpec extends ObjectBehavior
     }
 
     public function it_does_nothing_when_its_not_a_product(
-        Connection $connectionNotUsed,
+        TwoWayProductAssociationsSaver $productAssociationsSaver,
         GenericEvent $event,
         ProductModel $productModel
     ) {
-        $this->beConstructedWith($connectionNotUsed);
-
         $event->getSubject()->willReturn($productModel);
+
+        $productAssociationsSaver->saveInvertedAssociations(Argument::any(), Argument::any())->shouldNotBeCalled();
 
         $this->saveInvertedAssociations($event);
     }
 
     public function it_does_nothing_when_the_product_does_not_have_two_way_association(
-        Connection $connectionNotUsed,
+        ProductAssociationInterface $oneWayAssociation,
+        TwoWayProductAssociationsSaver $productAssociationsSaver,
         GenericEvent $event,
         Product $product,
-        ProductAssociation $productAssociation,
-        Collection $associationCollection,
-        AssociationType $associationType
+        Collection $associationCollection
     ) {
-        $this->beConstructedWith($connectionNotUsed);
-
-        $event->getSubject()->willReturn($product);
-        $associationCollection->toArray()->willReturn([$productAssociation]);
+        $associationCollection->toArray()->willReturn([$oneWayAssociation]);
         $product->getAssociations()->willReturn($associationCollection);
-        $productAssociation->getAssociationType()->willReturn($associationType);
-        $associationType->getId()->willReturn(1);
-        $associationType->isTwoWay()->willReturn(false);
+        $event->getSubject()->willReturn($product);
+
+        $productAssociationsSaver->saveInvertedAssociations(Argument::any(), Argument::any())->shouldNotBeCalled();
 
         $this->saveInvertedAssociations($event);
     }
 
     public function it_attempts_to_save_inverted_association_when_product_have_two_way_association(
-        Connection $connection,
+        TwoWayProductAssociationsSaver $productAssociationsSaver,
         GenericEvent $event,
-        ProductInterface $productWithTwoWayAssociation
+        ProductInterface $product,
+        ProductAssociationInterface $twoWayAssociation,
+        Collection $associationCollection
     ) {
-        $event->getSubject()->willReturn($productWithTwoWayAssociation);
+        $associationCollection->toArray()->willReturn([$twoWayAssociation]);
+        $product->getAssociations()->willReturn($associationCollection);
+        $event->getSubject()->willReturn($product);
 
-        $connection->beginTransaction()->shouldBeCalled();
-        $connection->executeUpdate(
-            Argument::any(),
-            ["association_type_ids" => [1], "owner_id" => 42],
-            ["association_type_ids" => Connection::PARAM_INT_ARRAY]
-        )->shouldBeCalled();
-
-        $connection->commit()->shouldBeCalled();
+        $productAssociationsSaver->saveInvertedAssociations($product, [$twoWayAssociation])->shouldBeCalled();
 
         $this->saveInvertedAssociations($event);
     }
 
-    public function it_rollback_and_throw_an_error_when_one_error_occurred(
-        Connection $connection,
+    public function it_attempts_to_save_only_inverted_two_way_association(
+        TwoWayProductAssociationsSaver $productAssociationsSaver,
         GenericEvent $event,
-        ProductInterface $productWithTwoWayAssociation
+        ProductInterface $product,
+        ProductAssociationInterface $twoWayAssociation,
+        ProductAssociationInterface $oneWayAssociation,
+        Collection $associationCollection
     ) {
-        $event->getSubject()->willReturn($productWithTwoWayAssociation);
+        $associationCollection->toArray()->willReturn([$oneWayAssociation, $twoWayAssociation]);
+        $product->getAssociations()->willReturn($associationCollection);
+        $event->getSubject()->willReturn($product);
 
-        $connection->beginTransaction()->shouldBeCalled();
-        $connection->executeUpdate(Argument::any(), Argument::any(), Argument::any())->willThrow('\Exception');
-        $connection->rollBack()->shouldBeCalled();
+        $productAssociationsSaver->saveInvertedAssociations($product, [$twoWayAssociation])->shouldBeCalled();
 
-        $this->shouldThrow('\Exception')->duringSaveInvertedAssociations($event);
+        $this->saveInvertedAssociations($event);
     }
 }
